Allow setting min/max TLS protocol version in libpq
In the backend there are GUCs to control the minimum and maximum TLS
versions to allow for a connection, but the clientside libpq lacked
this ability.  Disallowing servers which aren't providing secure TLS
protocols is of interest to clients, but we provide a maximum protocol
version setting by the same rationale as for the backend; to aid with
testing and to cope with misbehaving software.

This refactors the OpenSSL replacement functions for setting TLS
version from the backend to src/common to avoid code duplication.