Online enable and disable of data checksums
This allows data checksums to be enabled, or disabled, in a running
cluster without requiring a restart. The data checksum mode will be
set to "inprogress" while the checksumhelper background worker walks
the database and rewrites all the data pages to include checksums.
Once all pages have been checksummed, the data checksum mode turns
to "on". The synchronization of the checksumhelper is based on the
procbarrier infrastructure.

This adds a new section in the documention for checksums, covering
the basics on how to enable and disable with the various means that
are possible. More information can probably be added covering other
aspects of data checksums.

In passing, this patch removes the SAMPLE procbarrier and replaces
it with the CHECKSUM procbarrier, as there is now a consumer of the
procbarrier API and the "dummy" barrier can be removed (it was needed
to keep the enum with something).