Compatibility with OpenSSL 3.0.0
OpenSSL 3.0.0 deprecates a set of functions used in libpq:

SSL_CTX_load_verify_locations and X509_STORE_load_locations were
replaced with individual calls for file and directory, so replace
with the appropriate call since we only have configuration options
for files.

Usage of DH_check has been discouraged for a long time, and it was
finally deprecated with no direct replacement. Instead implement a
version using the EVP api which is the recommended alternative.
u