Refactor library specific SSL test setup into separate module
This moves TLS library specific setup and teardown to its own classes
under SSL::Server. If another TLS library is supported then this will
make it easier to plug into the existing testing framework.

This changeset was extracted from the larger patchset for supporting
libnss as a TLS backend in PostgreSQL.